style: reorganize code
